+++ public/index.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 require_once dirname(__DIR__) . '/vendor/autoload.php';
 
 use Cadasto\OpenEHR\MCP\Assistant\Apis\CkmClient;
-use Cadasto\OpenEHR\MCP\Assistant\Resources\Guidelines;
+use Cadasto\OpenEHR\MCP\Assistant\Resources\Guides;
 use Cadasto\OpenEHR\MCP\Assistant\Resources\Terminologies;
 use Mcp\Capability\Registry\Container;
 use Mcp\Schema\Enum\ProtocolVersion;
@@ -55,7 +55,7 @@ try {
 
     // Initialize API clients, resources, etc.
     $container->set(CkmClient::class, new CkmClient($logger));
-    $container->set(Guidelines::class, new Guidelines());
+    $container->set(Guides::class, new Guides());
     $container->set(Terminologies::class, new Terminologies());
 
     // Build the server
@@ -67,7 +67,7 @@ try {
         ->setContainer($container)
         ->setLogger($logger);
     // add resources
-    Guidelines::addResources($builder);
+    Guides::addResources($builder);
     Terminologies::addResources($builder);
 
     $server = $builder->build();